requirements.txt
pytorch-lightning
hydra-core
wandb
dvc
python-dotenv
tensorflow_datasets
rising
pytorch-lightning

# hydra-joblib-launcher